Initialize the repo and the folders
The package billy_ros_sharp contains files to interact with the VR headset and controlers
The launch file shows how to interact with ROS-sharp. The rviz configuration has been add to display the VR topics.
The python scripts are used to create an image from the data of the occupancy grid, translate it into an image and display it in the headset field of view (by using the plajected screen method to visualize video ros topic into the VR headset)